Deployment/Lambda/dcoders_service_dataset/requirements.txt
requests_toolbelt
Pillow
pandas